add routing with sample sign in and homepage pages
