+++ apps/mobile/package.json
@@ -9,6 +9,8 @@
     "react-native": "*",
     "react-native-config": "*",
     "react-native-svg": "*",
-    "@react-native-async-storage/async-storage": "*"
+    "@react-native-async-storage/async-storage": "*",
+    "react-router-native": "*",
+    "styled-components": "*"
   }
 }
