+++ apps/mobile/src/app/pages/homepage/homepage.spec.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { render } from '@testing-library/react-native';
+
+import Homepage from './homepage';
+
+describe('Homepage', () => {
+  it('should render successfully', () => {
+    const { container } = render(<Homepage />);
+    expect(container).toBeTruthy();
+  });
+});
