+++ apps/mobile/src/app/pages/page-signin/sign-in.spec.tsx
@@ -1,0 +1,11 @@
+import { render } from '@testing-library/react-native';
+
+import SignIn from './sign-in';
+
+describe('SignIn', () => {
+  it('should render successfully', () => {
+    const { container } = render(<SignIn />);
+    expect(container).toBeTruthy();
+  });
+});
+
