+++ package.json
@@ -13,9 +13,12 @@
     "@nestjs/core": "^9.0.0",
     "@nestjs/platform-express": "^9.0.0",
     "react": "18.2.0",
+    "react-is": "18.2.0",
     "react-native": "0.70.1",
+    "react-router-native": "^6.4.2",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.0.0",
+    "styled-components": "5.3.5",
     "tslib": "^2.3.0"
   },
   "devDependencies": {
@@ -41,10 +44,13 @@
     "@types/jest": "28.1.1",
     "@types/node": "16.11.7",
     "@types/react": "18.0.20",
+    "@types/react-is": "17.0.3",
     "@types/react-native": "0.70.3",
+    "@types/styled-components": "5.1.26",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
     "@typescript-eslint/parser": "^5.36.1",
     "babel-jest": "28.1.1",
+    "babel-plugin-styled-components": "1.10.7",
     "detox": "19.12.1",
     "eslint": "~8.15.0",
     "eslint-config-prettier": "8.1.0",
